Extracting the routerReducer so that it can be used directly in combineReducers (eg. fixes redux-loop)
