+++ src/createAll.js
@@ -1,6 +1,7 @@
 import * as actions from './actions'
 import createConnectedRouter from './ConnectedRouter'
-import createConnectRouter from './reducer'
+import createConnectRouter from './createConnectRouter'
+import createRouterReducer from './createRouterReducer'
 import routerMiddleware from './middleware'
 import createSelectors from './selectors'
 
@@ -9,6 +10,7 @@ const createAll = structure => ({
   ...createSelectors(structure),
   ConnectedRouter: createConnectedRouter(structure),
   connectRouter: createConnectRouter(structure),
+  routerReducer: createRouterReducer(structure),
   routerMiddleware,
 })
