+++ src/createConnectRouter.js
@@ -1,30 +1,21 @@
-import { LOCATION_CHANGE } from './actions'
+import createRouterReducer from './createRouterReducer'
 
 const createConnectRouter = (structure) => {
   const {
     filterNotRouter,
     fromJS,
     getIn,
-    merge,
     setIn,
   } = structure
-  /**
-   * This reducer will update the state with the most recent location history
-   * has transitioned to.
-   */
-  const routerReducer = (state, { type, payload } = {}) => {
-    if (type === LOCATION_CHANGE) {
-      return merge(state, payload)
-    }
-
-    return state
-  }
 
   const connectRouter = (history) => {
     const initialRouterState = fromJS({
       location: history.location,
       action: history.action,
     })
+
+    const routerReducer = createRouterReducer(structure)(history)
+
     // Wrap a root reducer and return a new root reducer with router state
     return (rootReducer) => (state, action) => {
       let routerState = initialRouterState
